edit code in school

+++ src/ft_strlcat.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strlcat.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: gtaggana <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/10/11 10:32:54 by gtaggana          #+#    #+#             */
+/*   Updated: 2021/10/11 10:32:55 by gtaggana         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
 size_t	ft_strlcat(char	*dst, const char *src, size_t size)
@@ -15,5 +27,5 @@ size_t	ft_strlcat(char	*dst, const char *src, size_t size)
 		size_dst++;
 	}
 	dst[size_dst] = '\0';
-	return (pt_dst_src + ft_strlen(src));
+	return (pt_dst_src + (size_t)ft_strlen(src));
 }
